+++ rollup.config.js
@@ -1,78 +1,36 @@
-import typescript from "@rollup/plugin-typescript"
 import nodePolyfills from "rollup-plugin-polyfill-node"
 import { nodeResolve } from "@rollup/plugin-node-resolve"
 import commonjs from "@rollup/plugin-commonjs"
 
 export default [
   {
-    input: "src/index.ts",
+    input: "dist/index.js",
     output: {
       format: "cjs",
       file: "dist/cjs/index.js",
       sourcemap: true,
     },
-    plugins: [
-      commonjs(),
-      nodeResolve(),
-      typescript({
-        module: "Node16",
-        moduleResolution: "Node16",
-        tsconfig: "./tsconfig.build.json",
-        compilerOptions: {
-          outDir: "dist/cjs",
-          declaration: true,
-          declarationDir: "dist/cjs",
-        },
-      }),
-      nodePolyfills(),
-    ],
+    external: ["rhea", "uuid"],
+    plugins: [commonjs(), nodePolyfills()],
   },
   {
-    input: "src/index.ts",
+    input: "dist/index.js",
     output: {
       format: "es",
       file: "dist/esm/index.js",
       name: "RabbitmqClient",
       sourcemap: true,
     },
-    plugins: [
-      commonjs(),
-      nodeResolve(),
-      typescript({
-        module: "Node16",
-        moduleResolution: "Node16",
-        tsconfig: "./tsconfig.build.json",
-        compilerOptions: {
-          outDir: "dist/esm",
-          declaration: true,
-          declarationDir: "dist/esm",
-        },
-      }),
-      nodePolyfills(),
-    ],
+    plugins: [commonjs(), nodeResolve(), nodePolyfills()],
   },
   {
-    input: "src/index.ts",
+    input: "dist/index.js",
     output: {
       format: "umd",
       file: "dist/umd/index.js",
       name: "RabbitmqClient",
       sourcemap: true,
     },
-    plugins: [
-      commonjs(),
-      nodeResolve(),
-      typescript({
-        module: "Node16",
-        moduleResolution: "Node16",
-        tsconfig: "./tsconfig.build.json",
-        compilerOptions: {
-          outDir: "dist/umd",
-          declaration: true,
-          declarationDir: "dist/umd",
-        },
-      }),
-      nodePolyfills(),
-    ],
+    plugins: [commonjs(), nodeResolve(), nodePolyfills()],
   },
 ]
