+++ src/message.ts
@@ -1,7 +1,6 @@
 import { generate_uuid, MessageAnnotations, Message as RheaMessage } from "rhea"
 import { AmqpEndpoints } from "./link_message_builder.js"
 import { inspect } from "util"
-import { CreateConsumerParams } from "./consumer.js"
 
 export type ExchangeOptions = {
   name: string
@@ -44,11 +43,4 @@ export function createPublisherAddressFrom(options?: DestinationOptions): string
   }
 
   throw new Error(`Unknown publisher options -- ${inspect(options)}`)
-}
-
-export function createConsumerAddressFrom(params: CreateConsumerParams): string | undefined {
-  if ("queue" in params) return `/${AmqpEndpoints.Queues}/${params.queue.name}`
-  if ("stream" in params) return `/${AmqpEndpoints.Queues}/${params.stream.name}`
-
-  throw new Error(`Unknown publisher options -- ${inspect(params)}`)
 }
